+++ packages/ragbits-agents/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add syntax sugar allowing easier Agents definition (#820)
 - Add post-processors (#821)
 - Support streaming from downstream agents (#812)
+- Add supervisor post-processor (#830)
 
 ## 1.3.0 (2025-09-11)
 ### Changed